Added Ajtai Commitment "working".
Additionaly:
1. Changed Context API to work with ring and not with
   modules
2. Changed Falcon files to reflect previous change.
3. Added simple test case for ajtai.

It simply tests that a commitment generated correctly
works. The hash method to point is a whacky hack that
somewhat should work, but there are such much better
ways. It simplifies by generating a random vector,
multiplying it [1, w] vector to generate a hash function
(i.e. using Ajtai One-Way Function). It seemed easier to
implement than going with ChaCha, or SHA digestures from
point.

For reference, if one should implement it for consumers,
you should use a similar function that the one provided
by the falcon sub-package.

Added Ajtai Commitment, but it does not work.

I need to change to vectors and make a few changes so it is
possible to create _hash_to_point_factory, or change the
implementation.

At the end of the day, it was made so that it is easy to compute
the challenge.

That being said, as we are working on Ring and not Module,
Context should reflect that.

Added thingy.

Finished Ajtai commitment